chore: rename userDetails to UserSignUpDetails to both entity and DTO and update UserSignUpDetailsDto tests
